+++ src/SharpLink.Abstractions/AssemblyInternals.cs
@@ -2,4 +2,5 @@ using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("SharpLink.Server")]
 [assembly: InternalsVisibleTo("SharpLink.Client")]
+[assembly: InternalsVisibleTo("SharpLink.Runtime")]
 [assembly: InternalsVisibleTo("SharpLink.UnitTests")]